backend/requirements.txt

# API-only dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Web framework and CORS
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# Desktop wrapper
pywebview>=4.4
pymongo>= 4.14.0

# Document processing (for resume upload)
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0 
requests

# Whisper for speech recognition
openai-whisper>=2024.04.10
